moved seeds files into db folder. modified styles for trending bar to have glow effect on images. added game_desc to Game Model to allow for a brief game description
