+++ db/schema.sql
@@ -1,6 +1,6 @@
 -- updated....sequelise will create the tables
 -- DROP DATABASE
-DROP DATABASE IF EXISTS ecommerce_db;
+DROP DATABASE IF EXISTS gamreview_db;
 
 -- CREATE DATABASE
-CREATE DATABASE ecommerce_db;
+CREATE DATABASE gamereview_db;
