+++ db/seeds/user-seeds.js
@@ -1,4 +1,4 @@
-const { User } = require('../models');
+const { User } = require('../../models');
 
 const userData = [
     {
