+++ views/layouts/index.html
@@ -42,36 +42,37 @@
 					<div class="trendingImg">
 						<img src="../../public/images/game-placeholder-image.jpg" />
 					</div>
-					<div class="trendingTitle">T I T L E</div>
-					<div class="trendingRating">R A T I N G</div>
+					<!-- Note: consider removing the trendingTitle div if game images will come with their titles default in the pic -->
+					<div class="trendingTitle"><a>T I T L E</a></div>
+					<!-- <div class="trendingRating">R A T I N G</div> -->
 				</div>
 				<div class="trendingGames">
 					<div class="trendingImg">
 						<img src="../../public/images/game-placeholder-image.jpg" />
 					</div>
-					<div class="trendingTitle">T I T L E</div>
-					<div class="trendingRating">R A T I N G</div>
+					<div class="trendingTitle"><a>T I T L E</a></div>
+					<!-- <div class="trendingRating">R A T I N G</div> -->
 				</div>
 				<div class="trendingGames">
 					<div class="trendingImg">
 						<img src="../../public/images/game-placeholder-image.jpg" />
 					</div>
-					<div class="trendingTitle">T I T L E</div>
-					<div class="trendingRating">R A T I N G</div>
+					<div class="trendingTitle"><a>T I T L E</a></div>
+					<!-- <div class="trendingRating">R A T I N G</div> -->
 				</div>
 				<div class="trendingGames">
 					<div class="trendingImg">
 						<img src="../../public/images/game-placeholder-image.jpg" />
 					</div>
-					<div class="trendingTitle">T I T L E</div>
-					<div class="trendingRating">R A T I N G</div>
+					<div class="trendingTitle"><a>T I T L E</a></div>
+					<!-- <div class="trendingRating">R A T I N G</div> -->
 				</div>
 				<div class="trendingGames">
 					<div class="trendingImg">
 						<img src="../../public/images/game-placeholder-image.jpg" />
 					</div>
-					<div class="trendingTitle">T I T L E</div>
-					<div class="trendingRating">R A T I N G</div>
+					<div class="trendingTitle"><a>T I T L E</a></div>
+					<!-- <div class="trendingRating">R A T I N G</div> -->
 				</div>
 			</div>
 		</div>
